+++ src/main/java/ru/yandex/Main.java
@@ -16,7 +16,7 @@ public class Main {
         System.out.println("Введите второе число");
         String secondArg = scanner.nextLine();
 
-        Integer result = calculator.calculate(Integer.parseInt(firstArg), Integer.parseInt(secondArg), action);
+        Double result = calculator.calculate(Double.parseDouble(firstArg), Integer.parseInt(secondArg), action);
 
         System.out.println("Результат вычислений: " + result);
     }
